Remove private topics. localgovdrupal/localgov_services#17
The field_all_topics should have the contents of the previous
field_topic_terms and the field_private_topics. This was maintained
in code when creating and updating the entity. If not the
field_topic_terms could just be mapped to the classified field,
or the contents of both need merging.