Add plugin to extract the geo address and geocode from legacy addressfield
BHCC had a custom address field that was used to store the geolocations.
This was not always correctly geocoded to the embeded geofield, which meant
that the geo from directory venue migration was getting skipped.
This will also try to geocode them with the localgov_default_osm if it can't
extracts a lat / lon pair.